+++ changelog.md
@@ -22,6 +22,7 @@ Changelog
 - Listing instances now shows indices also. (v4.2.0)
 - Some code style and documentation update. (v4.2.1)
 - Set a custom name for a project. See [Adding New Project](readme.md#add-new-project) section. 🗃️ (v4.2.2) 
+- Ability to rename a project. (v4.2.3)
 
 #### Version 3.x